// packages/services/relayer/src/rpc-relayer/relayer.gen.ts
export enum ETHTxnStatus {
  UNKNOWN = 'UNKNOWN',
  DROPPED = 'DROPPED',
  QUEUED = 'QUEUED',
  SENT = 'SENT',
  SUCCEEDED = 'SUCCEEDED',
  PARTIALLY_FAILED = 'PARTIALLY_FAILED',
  FAILED = 'FAILED',
  PENDING_PRECONDITION = 'PENDING_PRECONDITION'
}

export enum TransferType {
  SEND = 'SEND',
  RECEIVE = 'RECEIVE',
  BRIDGE_DEPOSIT = 'BRIDGE_DEPOSIT',
  BRIDGE_WITHDRAW = 'BRIDGE_WITHDRAW',
  BURN = 'BURN',
  UNKNOWN = 'UNKNOWN'
}

export enum SimulateStatus {
  SKIPPED = 'SKIPPED',
  SUCCEEDED = 'SUCCEEDED',
  FAILED = 'FAILED',
  ABORTED = 'ABORTED',
  REVERTED = 'REVERTED',
  NOT_ENOUGH_GAS = 'NOT_ENOUGH_GAS'
}

export enum FeeTokenType {
  UNKNOWN = 'UNKNOWN',
  ERC20_TOKEN = 'ERC20_TOKEN',
  ERC1155_TOKEN = 'ERC1155_TOKEN'
}
// ...
export interface MetaTxn {
  walletAddress: string
  contract: string
  input: string
}
// ...
export interface IntentPrecondition {
  type: string
  chainId: string
  data: any
}
// ...
export interface FeeToken {
  chainId: number
  name: string
  symbol: string
  type: FeeTokenType
  decimals?: number
  logoURL: string
  contractAddress?: string
  tokenID?: string
}
// ...
export interface SendMetaTxnReturn {
  status: boolean
  txnHash: string  
}